final-project/diffusion-based/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "diffusion-based-cifar10lt"
version = "1.0.0"
description = "Diffusion-based oversampling for handling imbalanced CIFAR-10-LT data"
authors = [{name = "ML Researcher", email = "researcher@example.com"}]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "datasets>=2.14.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.4.0",
    "flake8>=6.0.0",
    "jupyter>=1.0.0",
    "ipython>=8.12.0",
]

[tool.poetry]
name = "diffusion-based-cifar10lt"
version = "1.0.0"
description = "Diffusion-based oversampling for handling imbalanced CIFAR-10-LT data"
authors = ["ML Researcher <researcher@example.com>"]
readme = "README.md"
pakage-mode=false

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.0"
torchvision = "^0.15.0"
datasets = "^2.14.0"
pyyaml = "^6.0"
numpy = "^1.24.0"
scipy = "^1.10.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
tqdm = "^4.65.0"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.4.0"
flake8 = "^6.0.0"
jupyter = "^1.0.0"
ipython = "^8.12.0"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
